Move logic to a helper
